requirements.txt
aenum==3.1.11
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.0
apache-airflow==2.6.3
apache-airflow-providers-common-sql==1.6.0
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-google==10.4.0
apache-airflow-providers-http==4.5.0
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-sqlite==3.4.2
apispec==5.2.2
appdirs==1.4.4
APScheduler==3.9.1
argcomplete==3.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-lru==2.0.2
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.24.1
azure-storage-blob==12.12.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
bleach==6.0.0
blinker==1.6.2
bs4==0.0.1
cachelib==0.9.0
cachetools==5.3.0
case-conversion==2.1.0
cattrs==23.1.2
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.6
clickclick==20.10.2
clipboard==0.0.4
colorama==0.4.4
colorlog==4.8.0
colour==0.1.5
comm==0.1.3
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.2
cursor==1.3.5
db-dtypes==1.1.1
debugpy==1.6.7
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.5.1
dnspython==2.4.0
docutils==0.20.1
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
Flask==2.2.5
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
fuzzywuzzy==0.18.0
gcloud-aio-auth==4.2.3
gcloud-aio-bigquery==6.3.0
gcloud-aio-storage==8.3.0
gcsfs==2023.4.0
gender-guesser==0.4.0
google-ads==21.2.0
google-api-core==2.11.0
google-api-python-client==2.94.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-aiplatform==1.28.1
google-cloud-appengine-logging==1.3.1
google-cloud-audit-log==0.2.5
google-cloud-automl==2.11.2
google-cloud-bigquery==3.10.0
google-cloud-bigquery-datatransfer==3.11.2
google-cloud-bigquery-storage==2.22.0
google-cloud-bigtable==2.20.0
google-cloud-build==3.19.0
google-cloud-compute==1.13.0
google-cloud-container==2.27.0
google-cloud-core==2.3.2
google-cloud-datacatalog==3.14.0
google-cloud-dataflow-client==0.8.4
google-cloud-dataform==0.5.2
google-cloud-dataplex==1.5.1
google-cloud-dataproc==5.4.2
google-cloud-dataproc-metastore==1.12.0
google-cloud-dlp==3.12.2
google-cloud-kms==2.18.0
google-cloud-language==2.10.1
google-cloud-logging==3.6.0
google-cloud-memcache==1.7.2
google-cloud-monitoring==2.15.1
google-cloud-orchestration-airflow==1.9.1
google-cloud-os-login==2.10.0
google-cloud-pubsub==2.18.0
google-cloud-redis==2.13.1
google-cloud-resource-manager==1.10.2
google-cloud-secret-manager==2.16.2
google-cloud-spanner==3.38.0
google-cloud-speech==2.21.0
google-cloud-storage==2.8.0
google-cloud-storage-transfer==1.9.1
google-cloud-tasks==2.13.2
google-cloud-texttospeech==2.14.1
google-cloud-translate==3.11.2
google-cloud-videointelligence==2.11.3
google-cloud-vision==3.4.4
google-cloud-workflows==1.10.2
google-crc32c==1.5.0
google-re2==1.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
graphviz==0.20.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-gcp==0.2.2
grpcio-status==1.54.0
gunicorn==21.2.0
h11==0.14.0
html-text==0.5.2
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-resources==6.0.0
infi.systray==0.1.12
inflect==5.6.0
inflection==0.5.1
ipykernel==6.23.1
ipython==8.14.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
json-merge-patch==0.2
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.1
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
looker-sdk==23.12.0
lxml==4.9.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.6
maybe-else==0.2.1
mbstrdecoder==1.1.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==2.0.5
moviepy==1.0.3
msoffcrypto-tool==5.0.0
msrest==0.7.1
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook_shim==0.2.3
numpy==1.22.4
O365==2.0.19
oauthlib==3.2.0
office365==0.3.15
olefile==0.46
openpyxl==3.0.10
ordered-set==4.1.0
overrides==7.3.1
packaging==23.1
pandas==1.4.2
pandas-gbq==0.19.2
pandocfilters==1.5.0
parsedatetime==2.6
parso==0.8.3
pathmagic==0.3.14
pathspec==0.9.0
pendulum==2.1.2
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==3.5.1
pluggy==1.2.0
prettierfier==1.0.3
prison==0.2.1
proglog==0.1.10
prometheus-client==0.17.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.11
pydata-google-auth==1.8.1
pydub==0.25.1
Pygments==2.15.1
pyinstrument==4.1.1
pyiotools==0.3.18
PyJWT==2.8.0
pymiscutils==0.3.14
pyOpenSSL==23.2.0
pyparsing==3.1.0
PyPDF2==2.2.0
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyrsistent==0.19.3
pysubtypes==0.3.18
python-daemon==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-json-logger==2.0.7
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
readchar==3.0.5
regex==2022.6.2
requests==2.28.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rich-argparse==1.2.0
rsa==4.9
Send2Trash==1.8.0
setproctitle==1.3.2
Shapely==1.8.5.post1
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sql-formatter==0.6.2
SQLAlchemy==1.4.49
sqlalchemy-bigquery==1.7.0
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.2
stringcase==1.2.0
suds==1.1.2
tabulate==0.8.9
tenacity==8.2.2
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
tqdm==4.64.0
traitlets==5.9.0
typepy==1.3.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uc-micro-py==1.0.2
unicodecsv==0.14.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
XlsxWriter==3.0.3
xmlsec==1.3.13
yarl==1.9.1
